requirements.txt
google-generativeai
python-dotenv
pathlib